+++ uui-components/src/inputs/__tests__/TextInput.test.tsx
@@ -1,5 +1,7 @@
 import * as React from 'react';
-import { fireEvent, screen, setupComponentForTest, userEvent } from '@epam/uui-test-utils';
+import { fireEvent, renderWithContextAsync, screen, setupComponentForTest, userEvent } from '@epam/uui-test-utils';
+// @ts-ignore TODO: fix the TypeScript error about not being able to find the package.
+import { ReactComponent as CalendarIcon } from '@epam/assets/icons/action-calendar-fill.svg';
 import { TextInput, TextInputProps } from '../TextInput';
 
 jest.mock('@epam/uui-core', () => {
@@ -16,7 +18,7 @@ async function setupTextInput(params: Partial<TextInputProps>) {
         (context) => ({
             value: params.value,
             onValueChange: params.onValueChange || jest.fn().mockImplementation((newValue) => {
-                context.current.setProperty('value', newValue);
+                context.current?.setProperty('value', newValue);
             }),
             ...params,
         }),
@@ -31,17 +33,6 @@ async function setupTextInput(params: Partial<TextInputProps>) {
 }
 
 describe('TextInput', () => {
-    it('handles input change', async () => {
-        const handleChange = jest.fn();
-        await setupTextInput({ value: '', onValueChange: handleChange });
-        const input = screen.getByRole('textbox');
-
-        fireEvent.change(input, { target: { value: 'Test' } });
-
-        expect(handleChange).toHaveBeenCalledTimes(1);
-        expect(handleChange).toHaveBeenCalledWith('Test');
-    });
-
     it('handles input change', async () => {
         const handleChange = jest.fn();
         await setupTextInput({ value: '', onValueChange: handleChange });
@@ -95,24 +86,6 @@ describe('TextInput', () => {
         expect(handleClick).toHaveBeenCalledTimes(1);
     });
 
-    it('prevents default click behavior when target has uuiMarkers.clickable class', async () => {
-        const handleClick = jest.fn();
-        const handleChange = jest.fn();
-        await setupTextInput({
-            value: '',
-            onValueChange: handleChange,
-            onClick: handleClick,
-            rawProps: {
-                'data-testid': 'uui-text-input-wrapper',
-            },
-        });
-        const wrapper = screen.queryByTestId('uui-text-input-wrapper') as HTMLDivElement;
-
-        await userEvent.click(wrapper);
-
-        expect(handleClick).not.toHaveBeenCalled();
-    });
-
     it('handles key down event', async () => {
         const handleKeyDown = jest.fn();
         const handleAccept = jest.fn();
@@ -141,5 +114,88 @@ describe('TextInput', () => {
 
         expect(handleChange).toHaveBeenCalledTimes(1);
         expect(getValueChangeAnalyticsEvent).toHaveBeenCalled();
+    });
+
+    it('clears input via keyboard using clear button', async () => {
+        function TestComponent(): React.ReactNode {
+            const valueInitial = undefined;
+            const [
+                value,
+                setValue,
+            ] = React.useState<string | undefined>(valueInitial);
+
+            return (
+                <TextInput
+                    value={ value }
+                    onValueChange={ setValue }
+                    onCancel={ () => {
+                        setValue(valueInitial);
+                    } }
+                />
+            );
+        }
+
+        await renderWithContextAsync(
+            <TestComponent />,
+        );
+
+        // Initial state check (empty and unfocused input).
+        const input = await screen.findByRole('textbox');
+        expect(input).toHaveValue('');
+        expect(input).not.toHaveFocus();
+
+        // User clicks on the input to start typing.
+        await userEvent.click(input);
+        expect(input).toHaveFocus();
+
+        /*
+            User starts typing some text in the input,
+            which makes the clear button to appear.
+        */
+        await userEvent.type(
+            input,
+            'Test',
+        );
+        expect(input).toHaveValue('Test');
+        expect(input).toHaveFocus();
+        const clearButton = await screen.findByRole(
+            'button',
+            {
+                name: /clear input/i,
+            },
+        );
+        expect(clearButton).toBeInTheDocument();
+
+        // User changes focus to the clear button.
+        await userEvent.tab();
+        expect(input).not.toHaveFocus();
+        expect(clearButton).toHaveFocus();
+
+        // User presses "Enter" key to clear the input's value.
+        await userEvent.keyboard('{Enter}');
+        expect(input).toHaveValue('');
+        expect(input).toHaveFocus();
+        expect(clearButton).not.toBeInTheDocument();
+    });
+
+    it('calls onIconClick', async () => {
+        const handleChange = jest.fn();
+        const onIconClick = jest.fn();
+        await setupTextInput({
+            value: '',
+            onValueChange: handleChange,
+            icon: CalendarIcon,
+            onIconClick,
+        });
+        const clickableIcon = screen.getByRole(
+            'button',
+            {
+                name: /icon in input/i,
+            },
+        );
+
+        await userEvent.click(clickableIcon);
+
+        expect(onIconClick).toBeCalledTimes(1);
     });
 });